requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
altair==5.5.0
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
attrs==25.3.0
autogen-agentchat==0.5.7
autogen-core==0.5.7
autogen-ext==0.5.7
autogenstudio==0.4.2.2
azure-ai-agents==1.0.2
azure-ai-documentintelligence==1.0.2
azure-ai-inference==1.0.0b9
azure-ai-projects==1.0.0b12
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.23.1
azure-search-documents==11.5.3
azure-storage-blob==12.25.1
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
Chroma==0.2.0
chromadb==1.0.15
click==8.2.1
cobble==0.1.4
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
durationpy==0.10
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
groq==0.30.0
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
html2text==2025.4.15
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kubernetes==33.1.0
langchain==0.3.26
langchain-chroma==0.2.4
langchain-community==0.3.27
langchain-core==0.3.69
langchain-groq==0.3.6
langchain-huggingface==0.3.0
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langgraph==0.5.3
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.73
langsmith==0.4.6
loguru==0.7.3
lxml==6.0.0
magika==0.6.2
Mako==1.3.10
mammoth==1.9.1
markdown-it-py==3.0.0
markdownify==1.1.0
markitdown==0.1.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.47.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.22.1
openai==1.97.0
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.0
ormsgpack==1.10.0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pdfminer.six==20250506
pillow==11.3.0
playwright==1.53.0
posthog==5.4.0
propcache==0.3.2
protobuf==5.29.5
psycopg==3.2.9
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.2
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
sentry-sdk==2.33.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.47.1
streamlit==1.47.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
timm==1.0.17
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.2
triton==3.3.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
xlrd==2.0.2
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
youtube-transcript-api==1.0.3
zipp==3.23.0
zstandard==0.23.0
